requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
sqlalchemy[asyncio]==2.0.41
asyncpg==0.30.0
alembic==1.15.2
pydantic==2.11.3
pydantic-settings==2.9.1
python-jose[cryptography]==3.4.0
bcrypt==4.2.0
httpx==0.28.1
python-multipart==0.0.20
slowapi==0.1.9
authlib==1.6.0
itsdangerous==2.2.0
cloudinary==1.41.0
pytest==8.3.5
pytest-asyncio==0.25.3
